Update Sensor firmware to work with IM24 door sensors (CU-3644bwf) (#187)
Add an additional address to the BLE Scan Filter with the first 6 characters of IM24 door IDs hard coded in, allowing for the last 6 characters to be inputted using the Particle Console function.

All references to the IM21 in the code are changed to instead refer to a generic IM door sensor. Functions, comments and file names have been updated.

A new function, isDoorOpen(), replaces two macro definitions isDoorOpen and isDoorClosed, checking the door open bit instead of predefined values, to return whether the door is open or closed. 

A new function, isDoorStatusUnknown(), was added to ensure that the state machine remains in state 0 while the door sensor status is initialized to 0x99.