Do not transition to idle state upon receiving stillness or duration alert (CU-39w5av5) (CU-3c6my58) (#191)
This change modifies the sensor state machine to prevent it from transitioning to idle state after sending a stillness alert. Instead, it resets the stillness timer and transition back to stillness state. Additionally, this change also prevents the sensor state machine from transitioning to idle state after sending a duration alert. Instead, it transitions back to the duration state. The duration timer is not reset.