Add linter to Travis build process (CU-1k10ejd) (#190)
Added a new stage "firmware linting and formatting checks" to Travis CI flow. First, all .cpp and .h files in the src/ and test/ directories are checked for formatting errors using the clang-format-check.py script. If any formatting errors are found, the build fails. Otherwise, cppcheck is then run to check all files in the src/ directory for potential problems. If any errors are found, the build fails. Otherwise, the stage passes.

Added a script clang-format-all.py which formats all .cpp and .h files in the src/ and test/ directories as specified in the .clang-format file, and formatted all files accordingly.