Sentry log for sensor frequently missing door events
At present, the firmware keeps track of the previous 3 heartbeats
and whether or not they missed door events. If >1 of the last 3
heartbeats missed door events, doorMissedFreq will be set to true
in the JSON data sent to the Particle webhook.

vitals.js handles /api/heartbeat, and upon receiving this data
from the sensor, will check if doorMissedFreq is set to true;
if yes, a Sentry log will be sent.