Fixed response object of /api/message-clients to match conventions (#243)
Upon porting this feature to the BraveButtons repository, I discovered
that the response object returned by /api/message-clients does not
match the conventions described at the top of the api.js file in this
repository.

This PR does not change the functionality of this API request; it simply
moves some of the response object under a `data` object. Please read
the conventions information at the top of api.js for more information.