Replace Travis firmware linting step (#283)
Travis environment is depending on the mongodb-org which has an expired signing key. Luckily this error is only in the firmware linting step. As we're hoping to migrate to GitHub Actions for testing in the future, I'd rather replace this step with a suggestion that devs run linting scripts themselves in good faith, rather than wait for Travis to resolve this 😓 .